fix int

+++ doctor-event-api/src/main/java/io/terminus/doctor/event/dto/event/sow/DoctorFarrowingDto.java
@@ -46,7 +46,7 @@ public class DoctorFarrowingDto implements Serializable{
 
     private double birthNestAvg;    //出生窝重
 
-    private double dayAgeAvg;  // 平均日龄计算
+    private Integer dayAgeAvg;  // 平均日龄计算
 
     private Integer farrowingLiveCount; // 分娩猪的数量 = 活公猪 + 活母猪 = 健崽数量 + 弱仔数量 (拼窝后， 可能不相等， 差值 拼窝数量)
